requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-constraint>=1.4.0
typer>=0.12.0
rich>=13.0.0
pydantic>=2.6.0
numpy>=1.26.0